Refactor membership plan react components
