+++ client/src/components/subscriptions/subscriptionPageContainer.js
@@ -1,15 +1,18 @@
 import { connect } from 'react-redux';
 import SubscriptionPage from './subscriptionPage';
+import { fetchAllMembershipPlans } from '../../actions/membershipPlansActions';
 
 const mapStateToProps = state => {
     return {
-        currentUser: state.currentUser
+        currentUser: state.currentUser,
+        membershipPlans: state.membershipPlans,
+        subscriptions: state.subscriptions
     };
 };
 
 const mapDispatchToProps = dispatch => {
     return {
-        
+        fetchAllMembershipPlans: () => dispatch(fetchAllMembershipPlans())
     };
 };
